fix: revert file-summarizer to previous version
